pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "earnings_analyzer"
version = "1.0.2"
authors = [
    { name="Jeremiah Bohr", email="jeremiah.bohr@gmail.com" },
]
description = "A Python package to fetch, scrape, and analyze earnings call transcripts."
readme = "README.md"
requires-python = ">=3.9"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11.0",
    "sec-api>=1.0.0",
    "beautifulsoup4>=4.12.3,<5",
    "lxml>=5.2.2,<6",
    "googlesearch-python>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest >= 7.0",
    "black",
    "flake8"
]

[project.urls]
"Homepage" = "https://github.com/jeremiahbohr/earnings-analyzer"
"Bug Tracker" = "https://github.com/jeremiahbohr/earnings-analyzer/issues"

[project.scripts]
earnings-analyzer = "earnings_analyzer.main:main"

[tool.setuptools.packages.find]
where = ["src"]